requirements.txt
Flask==0.10.1
Flask-Login==0.2.9
Flask-Mail==0.9.0
Flask-Restless==0.12.1
Flask-SQLAlchemy==1.0
Jinja2==2.7.2
MarkupSafe==0.18
SQLAlchemy==0.9.2
Werkzeug==0.9.4
argparse==1.1
blinker==1.3
gunicorn==18.0
heroku==0.1.4
httplib2==0.8
itsdangerous==0.23
passlib==1.6.2
phonenumbers==6.0.0
psycopg2==2.5.2
pycrypto==2.6.1
python-dateutil==1.5
pytz==2014.1.1
requests==2.2.1
six==1.5.2
tropo-webapi-python==0.1.3
twilio==3.6.5
wsgiref==0.1.2
yolk==0.4.3
coverage==3.7.1
